MDL-40243 Suspend users in addition to disabling them
Assume that we configure Authentication > LDAP > cron synchronization
script > Removed ext user "Suspend internal".

When we remove a user from LDAP directory and run the sync_users
script manually, the user doesn't appear as suspended in the user
management UI. The reason is the LDAP auth plugin doesn't suspend
users, it disables them (which is a subtle difference). It simply
assigns the user the 'nologin' auth plugin, so she isn't able to log
in (but otherwise doesn't modify the user at all).

As the user edition UI takes into account the 'suspended' property and
the LDAP plugin doesn't modify it, the user doesn't appear as
suspended. We need to change the 'suspended' property in addition to
assigning the 'nologin' auth plugin (and reverse things back when
reviving the user).