MDL-58395 ldap: sync should give problematic username
When the LDAP auth sync task is run, and it fails to create a new user
because some of the user details are problematic, current code doesn't
show the problematic username. In fact, doesn't show any user detail
at all. The last name that is shown in the task execution output is
the last known good user, not the problematic one. So it is very hard
to know which user is the problematic one and fixit.

This is espcially problematic, as the whole user addition process runs
in a transaction. And if it fails to create a single user, all the
users created in the current sync run are rolled back (and re-added
next run, probably to be rolled back again if the problematic user is
still un-fixed).

So give as much details about the user and the error type as possible,
and stop the sync process in a controlled way (instead of blowing the
whole thing from the exception)